feat: add an initial publiccode.yml
This PR adds a basic publiccode.yml file.
It also add a linter check to the codequality script,
and a corresponding ci-check.

publiccode.yml is a metadata specification,
read more about it here https://yml.publiccode.tools/

The publiccode.yml added here is to be seen as basic, not final,
and will change in the future, when having discussed a few
topics.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>